Move changes from /ansible_base to /lib
rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED
