+++ ansible_base/lib/workload_identity/base.py
