+++ test_app/tests/lib/workload_identity/test_controller.py
@@ -1,4 +1,4 @@
-from ansible_base.workload_identity.controller import AutomationControllerJobScope
+from ansible_base.lib.workload_identity import AutomationControllerJobScope
 
 
 def test_list_claims():
